add repo Interfaces. Changed hibernate fields and add TitleType. DPJ the service with the repositorys

+++ mywatchlist/src/main/java/mywatchlist/model/hibernate/UserAccount.java
@@ -9,7 +9,7 @@ public class UserAccount {
     @Id
     @GeneratedValue
     @Column(name = "user_id", nullable = false)
-    private int userId;
+    private Long userId;
     @Column(name = "username", nullable = false)
     private String username;
     @Column(name = "password", nullable = false)
@@ -19,11 +19,11 @@ public class UserAccount {
     @Column(name = "private_profile")
     private boolean privateProfile;
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
